A dual function bird feeder with a hex grid base
The shape of the top fits on top of a fence post so that it doesn't slip
down.  The base has a cool grid shape design that'll save a lot of
filament.  It's based on the dimensions of a wooden bird feeder that's
rotting on the fence.  Because the base is full of holes it probably
won't be so useful for holding bird seed.  Maybe drinks then...