New design: GOAT Tool fork
A three-pronged fork shape based for use on the GOAT multitool, for
eating lunch snacks on the go.  The GOAT project provides a tool
template file that I imported into OpenSCAD and cut the fork prongs out
of.  The fork shape itself seems to work well for eating but the PETG
material I printed it from might not be the strongest because it broke
at the hinge on the first day.  In future it'd also be cool to make the
design more parametric so that people can use the online "Customizer"
tool to adjust the number and length of the prongs etc.