Improve lamp strengthener design internals and adjust sizes
The previous one didn't fit and I got some more accurate measurements.
Updating them in the file was a bit annoying because most of them were
just written in and it was hard to work out / remember which numbers
were for what, so I add more variables to make it more parametric.  The
two sizes I had to adjust (width of the holes) are new parameters and
the thickness (8 looks good as a default) is added onto that to set the
thickness of the walls.