New design: Spotlight Handle
My brother-in-law has a car battery powered spotlight, the battery sits
in a custom made box with metal handles welded on but it's pretty heavy
and cuts into your hands when you lift it.  This design is for a pair of
comfortable plastic handles that clip on to make the battery pack easier
to carry.